.history/skin_cancer_app/requirements_20250401005920.txt
streamlit==1.30.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.3
pillow==10.1.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
matplotlib==3.8.2
streamlit-option-menu==0.3.6
streamlit-chat==0.1.1
sqlite3==2.6.0